Change "~" to "^" for all dependencies and add `package-lock.json`
Projects that depended on json-proxy would, up until now,
get a lot of errors when running `npm-audit`, due to outdated
versions of morgan and express. To fix these errors and avoid
some similar errors in the future, this commit changes all tilde (~)
specifiers to caret (^) specifiers. This allows the most recent
version of a dependency with the same major version number to be
installed, not just the most recent minor version:
https://docs.npmjs.com/files/package.json#dependencies

The commit also adds a `package-lock.json`-file, which since
npm v5.0.0 has been used to keep transitive dependency
versions in sync between developers:
https://blog.npmjs.org/post/161081169345/v500

Tests: Ran `test` and `jshint` scripts with no errors.

One minor detail: My version version of npm (v6.5.0)
automatically sorts the entries in dependencies. This makes
the changes in `package.json` slightly hard to track, but
it had to be done sooner or later.